+++ nodejs/proto.sh
@@ -1,11 +1,11 @@
 #! /bin/bash
-rm -rf ./src
+rm -rf ./src/clientchannel
 
 cd ..
 
 protoc \
 	--plugin=./nodejs/node_modules/.bin/protoc-gen-ts_proto \
-	--ts_proto_out=./nodejs \
+	--ts_proto_out=./nodejs/src \
 	--ts_proto_opt=esModuleInterop=true \
 	--ts_proto_opt=env=node \
 	--ts_proto_opt=oneof=unions \