+++ nodejs/src/index.ts
@@ -1,0 +1,41 @@
+export { ServiceClient, ServiceServer, ServiceService } from "./clientchannel/service";
+
+export {
+    DeepPartial,
+    Exact,
+    Response,
+    Response_InitAck,
+    Response_InitNotAck,
+    Response_InvalidateGdprAck,
+    Response_InvalidateGdprNotAck,
+    Response_PublishAck,
+    Response_PublishNotAck,
+    Response_SnapshotAck,
+    Response_SnapshotNotAck,
+    Response_SubscribeAck,
+    Response_SubscribeNotAck,
+} from "./clientchannel/response";
+
+export {
+    Request,
+    Request_EventAck,
+    Request_EventNotAck,
+    Request_InitAction,
+    Request_InvalidateGdprAction,
+    Request_PublishAction,
+    Request_SnapshotAction,
+    Request_SubscribeAction,
+} from "./clientchannel/request";
+
+export { GetStreamRequest, GetStreamResponse } from "./clientchannel/get_stream";
+
+export { GetEventsFromStartRequest } from "./clientchannel/get_events_from_start";
+
+export {
+    Data,
+    Data_GdprMetadata,
+    Event,
+    EventEnvelope,
+    Event_PayloadEntry,
+    PublishEventEnvelope,
+} from "./clientchannel/event";